requirements.txt
autopep8
certifi
charset-normalizer
click
et-xmlfile
Flask
idna
itsdangerous
Jinja2
MarkupSafe
numpy
openai
openpyxl
pandas
pandas-stubs
pycodestyle
python-dateutil
python-dotenv
pytz
requests
six
toml
tqdm
urllib3
Werkzeug
wavio
sounddevice
scipy
